drones/requirements.txt
# Ainova Drone System — Python Dependencies
# RunPod: Python 3.11 + CUDA 12.4

# LLM serving
vllm>=0.6.0

# Supabase client
supabase>=2.0.0

# Web scraping
playwright>=1.40.0
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# HTTP
aiohttp>=3.9.0
requests>=2.31.0
httpx>=0.27.0

# Data processing
pydantic>=2.5.0

# CLI & logging
rich>=13.7.0
click>=8.1.0

# Utilities
python-dotenv>=1.0.0
